+++ src/util/index.ts
@@ -4,3 +4,4 @@ export * from "./runTask";
 export * from "./toAbsolute";
 export * from "./wait";
 export * from "./waitForDebugPort";
+export * from "./showErrorMessage";