+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Run editor contents command does not proceed if build/run tasks fail
 - Problem matchers correctly match absolute file paths
+- Display error messages from commands and tasks
 
 ## [1.8.0] - 2022-11-21